refactor: split the GitHub App routes by resource
routes/github-app.ts was 932 lines covering three unrelated resources
behind one Hono instance. Now a directory, one module per resource:

- installations.ts   install handshake + the rate-limited read-only
                     repository/environment proxies
- release-targets.ts the repo/environment pairs Drydock gates
- workflow-gates.ts  reviewer decisions, incl. the TOTP step-up path
- shared.ts          org-scoping check and error translation used by more
                     than one group

There is no router-level middleware — each handler establishes its own
organization context — so the groups compose without ordering constraints.
All 12 route paths are unchanged and each serializer moved with the
resource it serializes.
